requirements.txt
fastapi==0.72.0
uvicorn==0.17.0
rembg==2.0.13
Pillow==9.0.1
validators==0.20.0

